opt: move geo functions to a new file in xform
This commit does some refactoring by moving several geospatial functions
from xform/custom_funcs.go to a new file xform/geo.go. This is in
preparation for code changes in the next commit as well as more additions
to geo.go.

Release note: None
